Add White Text Style and Enhance App Drawer
Added a new TextStyle `bold18White` to the `AppTextStyles` class for white-colored, bold text with a font size of 18. This style is intended for use in the app's drawer where contrast against darker backgrounds is necessary.

Enhanced the `AppDrawer` widget to include a structured menu with options for "تعلم الحروف" (Learn Letters), "مارس ما تعلمته" (Practice What You Learned), "اختار الشخصية" (Choose a Character), and "اسأل سؤال" (Ask a Question). Each menu item is styled with the newly added `bold18White` text style and includes an icon colored with `AppColors.secondaryColor`. The drawer's layout is improved with spacing provided by the `Gap` widget for better visual separation.

Additionally, adjusted the horizontal padding of the `ElevatedButton` in the `PractiseViewBody` widget to 10 from 20 for better alignment and spacing consistency within the practice view.